Engine: Moved related files to folders
