+++ Libraries/Engine/EngineStandard.hpp
@@ -46,7 +46,7 @@ private:
 } // namespace Zero
 
 #include "Rectangle.hpp"
-#include "Resource.hpp"
+#include "Resource/Resource.hpp"
 #include "EngineBindingExtensions.hpp"
 #include "EngineObject.hpp"
 #include "EngineContainers.hpp"
@@ -60,8 +60,8 @@ private:
 #include "CogMeta.hpp"
 #include "Space.hpp"
 #include "DocumentResource.hpp"
-#include "ZilchResource.hpp"
-#include "ResourceLibrary.hpp"
+#include "Scripting/ZilchResource.hpp"
+#include "Resource/ResourceLibrary.hpp"
 #include "JobSystem.hpp"
 #include "EngineEvents.hpp"
 #include "System.hpp"
@@ -71,19 +71,19 @@ private:
 #include "Game.hpp"
 #include "Factory.hpp"
 #include "ArchetypeRebuilder.hpp"
-#include "ResourceSystem.hpp"
-#include "ResourcePropertyOperations.hpp"
+#include "Resource/ResourceSystem.hpp"
+#include "Resource/ResourcePropertyOperations.hpp"
 #include "ErrorContext.hpp"
 #include "Tracker.hpp"
 #include "Hierarchy.hpp"
 #include "TransformSupport.hpp"
 #include "Transform.hpp"
-#include "Action.hpp"
-#include "ActionSystem.hpp"
-#include "ActionEase.hpp"
-#include "MetaAction.hpp"
-#include "BasicActions.hpp"
-#include "ActionGenerator.hpp"
+#include "Action/Action.hpp"
+#include "Action/ActionSystem.hpp"
+#include "Action/ActionEase.hpp"
+#include "Action/MetaAction.hpp"
+#include "Action/BasicActions.hpp"
+#include "Action/ActionGenerator.hpp"
 #include "EditorSupport.hpp"
 #include "CogSerialization.hpp"
 #include "ObjectLoader.hpp"
@@ -91,30 +91,30 @@ private:
 #include "CogInitializer.hpp"
 #include "CogPath.hpp"
 #include "ObjectLink.hpp"
-#include "Keyboard.hpp"
+#include "Platform/Input/Keyboard.hpp"
 #include "Environment.hpp"
 #include "SystemObjectManager.hpp"
-#include "OsWindow.hpp"
-#include "OsShell.hpp"
-#include "ResourceManager.hpp"
+#include "Platform/OsWindow.hpp"
+#include "Platform/OsShell.hpp"
+#include "Resource/ResourceManager.hpp"
 #include "Archetype.hpp"
-#include "Mouse.hpp"
+#include "Platform/Input/Mouse.hpp"
 #include "Level.hpp"
 #include "Operation.hpp"
-#include "ResourceListOperation.hpp"
+#include "Resource/ResourceListOperation.hpp"
 #include "MetaOperations.hpp"
 #include "CogRestoreState.hpp"
 #include "CogOperations.hpp"
-#include "AnimationNode.hpp"
-#include "PropertyTrack.hpp"
-#include "AnimationGraph.hpp"
-#include "AnimationGraphEvents.hpp"
-#include "Animation.hpp"
+#include "Animation/AnimationNode.hpp"
+#include "Animation/PropertyTrack.hpp"
+#include "Animation/AnimationGraph.hpp"
+#include "Animation/AnimationGraphEvents.hpp"
+#include "Animation/Animation.hpp"
 #include "CogSelection.hpp"
 #include "Configuration.hpp"
 #include "LauncherConfiguration.hpp"
 #include "DebugDraw.hpp"
-#include "ScriptEvents.hpp"
+#include "Scripting/ScriptEvents.hpp"
 #include "TextResource.hpp"
 #include "EditorSettings.hpp"
 #include "Documentation.hpp"
@@ -134,15 +134,15 @@ private:
 #include "Project.hpp"
 #include "RaycastProvider.hpp"
 #include "SimpleResourceFactory.hpp"
-#include "GamePadSystem.hpp"
-#include "JoystickSystem.hpp"
+#include "Platform/Input/GamePadSystem.hpp"
+#include "Platform/Input/JoystickSystem.hpp"
 #include "EventDirectoryWatcher.hpp"
 #include "CogRange.hpp"
 #include "CogHelpers.hpp"
 #include "ComponentHierarchy.hpp"
 #include "DataSource.hpp"
 #include "EngineLibraryExtensions.hpp"
-#include "ZilchManager.hpp"
+#include "Scripting/ZilchManager.hpp"
 #include "Spline.hpp"
 #include "HierarchySpline.hpp"
 #include "AsyncProcess.hpp"
