+++ Libraries/Engine/Precompiled.hpp
@@ -1,4 +1,4 @@
 // MIT Licensed (see LICENSE.md).
 #pragma once
 #include "EngineStandard.hpp"
-#include "ZilchAction.hpp"
+#include "Scripting/ZilchAction.hpp"
