+++ Libraries/Engine/Resource/Resource.cpp
