+++ Libraries/Engine/Resource/Resource.hpp
