+++ Libraries/Engine/Resource/ResourcePropertyOperations.cpp
