+++ Libraries/Engine/Resource/ResourcePropertyOperations.hpp
