+++ Libraries/Engine/Resource/ResourceSystem.hpp
