+++ Libraries/Engine/Scripting/ZilchResource.cpp
